vp-deliverkey/requirements.txt
ssl
smtplib
google-cloud-secret-manager
google-cloud-firestore